Added operate and modified the config file
Modified config_generator file such that it takes the server
name and returns dict with names as key and values as ip address.
Modifying existing hosts and config file is little tricky and
needs more code.

Also added the operate file if a server is deleted or number of
servers required is equal to the servers active case is handled.